32 accuracy replaced to 64; refactroring

+++ course/src/light.rs
@@ -4,6 +4,6 @@ use crate::vec3::Vec3;
 #[derive(Debug)]
 pub struct Light {
     pub position: Vec3,
-    pub intensity: f32,
-    pub color_distribution: HashMap<i32, f32> // wavelengths (nm), density
+    pub intensity: f64,
+    pub color_distribution: HashMap<i64, f64> // wavelengths (nm), density
 }
