+++ course/src/material.rs
@@ -4,9 +4,9 @@ use lazy_static::lazy_static;
 
 #[derive(Debug)]
 pub struct Material {
-    pub specular_reflection: f32,
-    pub transparency: f32,
-    pub diffuse_reflection: HashMap<i32, f32> // wavelength (nm), Kd
+    pub specular_reflection: f64,
+    pub transparency: f64,
+    pub diffuse_reflection: HashMap<i64, f64> // wavelength (nm), Kd
 }
 
 lazy_static! {
